renamed profile-form + restructured component + added styles
