+++ src/app/routes/register/components/meeting-place-form/meeting-place-form.component.html
@@ -20,41 +20,45 @@
 @if (isAddButtonClicked) {
   <div [formGroup]="preferredMeetingPlaceFormGroup" id="preferred-meeting-places">
     <strong class="number">{{ preferredMeetingPlacesDisplay.length + 1 }}</strong>
-    <div class="row my-3 mt-5">
-      <div class="col-8">
+    <div class="row justify-content-center mt-5">
+      <div class="col-12 col-md-8">
         <label for="name">Nom</label>
         <input type="text" formControlName="name" id="name" name="name" class="form-control" placeholder="Nom du lieu" required/>
         @if (preferredMeetingPlaceFormGroup.get('name')?.errors && (preferredMeetingPlaceFormGroup.get('name')?.dirty || preferredMeetingPlaceFormGroup.get('name')?.touched)) {
-          <div [hidden]="!preferredMeetingPlaceFormGroup.get('name')?.errors?.['required']" class="error-msg">Veuillez renseigner un nom de lieu</div>
-          <div [hidden]="!preferredMeetingPlaceFormGroup.get('name')?.errors?.['alreadyExists']" class="error-msg">Ce nom de lieu existe déjà</div>
+          <div class="error-msg mt-2">
+            <div [hidden]="!preferredMeetingPlaceFormGroup.get('name')?.errors?.['required']">Veuillez renseigner un nom de lieu</div>
+            <div [hidden]="!preferredMeetingPlaceFormGroup.get('name')?.errors?.['alreadyExists']">Ce nom de lieu existe déjà</div>
+          </div>
         }
       </div>
     </div>
-    <div class="row my-3 mt-5">
-      <div class="col-8">
+    <div class="row justify-content-center mt-5">
+      <div class="col-12 col-md-8">
         <label for="street">Rue</label>
         <mapbox-address-autofill>
           <input type="text" formControlName="street" autocomplete="address-line1" id="street" name="street" class="form-control" placeholder="Numéro et nom de rue"required/>
         </mapbox-address-autofill>
         @if (preferredMeetingPlaceFormGroup.get('street')?.errors && (preferredMeetingPlaceFormGroup.get('street')?.dirty || preferredMeetingPlaceFormGroup.get('street')?.touched)) {
-          <div [hidden]="!preferredMeetingPlaceFormGroup.get('street')?.errors?.['required']" class="error-msg">Veuillez renseigner une rue</div>
-          <div [hidden]="!preferredMeetingPlaceFormGroup.get('street')?.errors?.['alreadyExists']" class="error-msg">Cette adresse a déjà été enregistrée</div>
+          <div class="error-msg mt-2">
+            <div [hidden]="!preferredMeetingPlaceFormGroup.get('street')?.errors?.['required']">Veuillez renseigner une rue</div>
+            <div [hidden]="!preferredMeetingPlaceFormGroup.get('street')?.errors?.['alreadyExists']">Cette adresse a déjà été enregistrée</div>
+          </div>
         }
       </div>
     </div>
-    <div class="row my-3 mt-5">
-      <div class="col-4">
+    <div class="row justify-content-center mt-5">
+      <div class="col-6 col-md-4">
         <label for="postal-code">Code postal</label>
         <input type="text" formControlName="postalCode" autocomplete="postal-code" id="postal-code" name="postal-code" class="form-control" required/>
         @if (preferredMeetingPlaceFormGroup.get('postalCode')?.errors && (preferredMeetingPlaceFormGroup.get('postalCode')?.dirty || preferredMeetingPlaceFormGroup.get('postalCode')?.touched)) {
-          <div class="error-msg">Veuillez renseigner un code postal</div>
+          <div class="error-msg mt-2">Veuillez renseigner un code postal</div>
         }
       </div>
-      <div class="col-4">
+      <div class="col-6 col-md-4">
         <label for="city">Ville</label>
         <input type="text" formControlName="city" autocomplete="address-level2" id="city" name="city" class="form-control" required/>
         @if(preferredMeetingPlaceFormGroup.get('city')?.errors && (preferredMeetingPlaceFormGroup.get('city')?.dirty || preferredMeetingPlaceFormGroup.get('city')?.touched)) {
-          <div class="error-msg">Veuillez renseigner un code postal</div>
+          <div class="error-msg mt-2">Veuillez renseigner un code postal</div>
         }
       </div>
     </div>
