--- src/app/routes/register/profile-form/profile-form.component.scss
@@ -1,11 +1,0 @@
-@import 'color-variables';
-
-// TO DO : ajouter certains elements dans le style.scss ? et les supprimer de ad-form
-.add-icon, .remove-icon {
-  width: 27.25px;
-  height: 27.25px;
-}
-
-.add-text {
-  color: rgba(0, 0, 0, 0.38);
-}
