+++ src/app/routes/register/register.component.scss
@@ -1,0 +1,21 @@
+@import 'color-variables';
+
+#bank-account-form {
+  margin: 0 auto;
+  width: fit-content;
+}
+
+#profile-picture {
+  margin: 0 auto;
+  width: fit-content;
+}
+
+// TO DO : ajouter certains elements dans le style.scss ? et les supprimer de ad-form
+.add-icon, .remove-icon {
+  width: 27.25px;
+  height: 27.25px;
+}
+
+.add-text {
+  color: rgba(0, 0, 0, 0.38);
+}
